# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: open-kraken HTTP API
  version: 0.1.0
  description: |
    Browser-facing HTTP contract for the open-kraken migration workspace.
    WebSocket handshake and event schemas are documented in x-websocket and in
    docs/api/http-websocket-contract.md.
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspaceId}/conversations/{conversationId}/messages:
    get:
      operationId: listConversationMessages
      summary: List conversation messages
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/ConversationId'
        - name: before
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Message page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePageResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      operationId: createConversationMessage
      summary: Create a message in a conversation
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '200':
          description: Persisted message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMessageResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/v1/skills:
    get:
      operationId: listSkills
      summary: List filesystem skill catalog entries
      responses:
        '200':
          description: Skill catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillListResponse'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/skills/reload:
    post:
      operationId: reloadSkills
      summary: Rescan the filesystem skill catalog
      responses:
        '200':
          description: Reload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillReloadResponse'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/v1/tokens/events:
    post:
      operationId: recordTokenEvent
      summary: Record a token usage event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenEvent'
      responses:
        '201':
          description: Recorded token event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEvent'
        '400':
          $ref: '#/components/responses/Error400'
  /api/v1/tokens/stats:
    get:
      operationId: getTokenStats
      summary: Get aggregate token usage statistics
      responses:
        '200':
          description: Token statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenStats'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/tokens/activity:
    get:
      operationId: listTokenActivity
      summary: List recent token usage events
      responses:
        '200':
          description: Recent token events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenActivityResponse'
        '500':
          $ref: '#/components/responses/Error500'
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error:
                  code: auth.unauthorized
                  message: bearer token is missing or invalid
                  status: 401
                  requestId: req_01KN9B00000000000000000003
                  retryable: false
                  details:
                    workspaceId: ws_123
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbiddenScope:
              value:
                error:
                  code: auth.capability_denied
                  message: member cannot dispatch to this terminal
                  status: 403
                  requestId: req_01KN9B00000000000000000004
                  retryable: false
                  details:
                    workspaceId: ws_123
                    memberId: member_member
                    requiredCapability: terminal.dispatch
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error422:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message, status, requestId, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: integer
        requestId:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    MessageContent:
      type: object
      required: [type]
      properties:
        type:
          type: string
        text:
          type: string
    MessageAttachment:
      type: object
      required: [type]
      properties:
        type:
          type: string
        title:
          type: string
        roadmapId:
          type: string
        taskId:
          type: string
    Message:
      type: object
      required: [id, content, createdAt, isAi, status]
      properties:
        id:
          type: string
        senderId:
          type: string
        content:
          $ref: '#/components/schemas/MessageContent'
        createdAt:
          type: integer
          format: int64
        isAi:
          type: boolean
        status:
          $ref: '#/components/schemas/MessageStatus'
        attachment:
          $ref: '#/components/schemas/MessageAttachment'
    MessageStatus:
      type: string
      enum: [pending, sent, failed]
      description: |
        pending = persisted but downstream delivery or side effects still settling.
        sent = persisted and accepted by the current downstream path.
        failed = persisted but downstream delivery or side effects failed.
...
    MessagePageResponse:
      type: object
      required: [conversationId, items]
      properties:
        conversationId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        nextBefore:
          type: string
    CreateMessageRequest:
      type: object
      required: [clientMessageId, content]
      properties:
        clientMessageId:
          type: string
        content:
          $ref: '#/components/schemas/MessageContent'
        attachment:
          $ref: '#/components/schemas/MessageAttachment'
    CreateMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          $ref: '#/components/schemas/Message'
...
    Skill:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        path:
          type: string
        category:
          type: string
        contentSummary:
          type: string
    SkillListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    SkillReloadResponse:
      type: object
      required: [loaded, skipped, reloadedAt]
      properties:
        loaded:
          type: integer
        skipped:
          type: integer
        reloadedAt:
          type: string
          format: date-time
...
    TokenEvent:
      type: object
      required: [memberId, model, inputTokens, outputTokens, cost]
      properties:
        id:
          type: string
        memberId:
          type: string
        nodeId:
          type: string
        model:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        cost:
          type: number
        timestamp:
          type: string
          format: date-time
    TokenStats:
      type: object
      required: [scope, inputTokens, outputTokens, totalTokens, totalCost, eventCount]
      properties:
        scope:
          type: string
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        totalTokens:
          type: integer
        totalCost:
          type: number
        eventCount:
          type: integer
    TokenActivityResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TokenEvent'
        total:
          type: integer
...